+++ src/components/20-molecules/datepicker/react/DemoDatepickerOnDateChange.jsx
@@ -26,10 +26,10 @@ const DemoDatepickerOnDateChange = () => {
   // N.B. onBlur={undefined} is there to demonstrate undefined property values don't throw exceptions (cf. issue #1613)
   return (
     <div>
-      Selected date: {selectedDateString}
-      <br />
+      <p>Selected date: {selectedDateString}</p>
       <AXADatepickerReact
         inputfield
+        width="300px"
         id="datepicker-react-inputfield-on-date-change"
         data-test-id="datepicker-react-inputfield-on-date-change"
         locale={locale}